+++ packages/core/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @stellar-agent-kit/core
 
+## 0.1.5
+
 ## 0.1.4
 
 ### Patch Changes